+++ app/src/main/java/com/example/ideo7/weather/Model/Main.java
@@ -6,7 +6,7 @@ package com.example.ideo7.weather.Model;
 
 public class Main {
     private Double temp;
-    private Integer pressure;
+    private Double pressure;
     private Integer humidity;
     private Double tempMin;
     private Double tempMax;
@@ -19,11 +19,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
-    public void setPressure(Integer pressure) {
+    public void setPressure(Double pressure) {
         this.pressure = pressure;
     }
 
@@ -54,7 +54,7 @@ public class Main {
     public Main() {
     }
 
-    public Main(Double temp, Integer pressure, Integer humidity, Double tempMin, Double tempMax) {
+    public Main(Double temp, Double pressure, Integer humidity, Double tempMin, Double tempMax) {
         this.temp = temp;
         this.pressure = pressure;
         this.humidity = humidity;
