+++ app/src/main/java/com/example/ideo7/weather/Model/Wind.java
@@ -6,7 +6,7 @@ package com.example.ideo7.weather.Model;
 
 public class Wind {
     private Double speed;
-    private Integer deg;
+    private Double deg;
 
     public Double getSpeed() {
         return speed;
@@ -16,18 +16,18 @@ public class Wind {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
 
     public Wind() {
     }
 
-    public Wind(Double speed, Integer deg) {
+    public Wind(Double speed, Double deg) {
         this.speed = speed;
         this.deg = deg;
     }
